implement authorize using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -7,6 +7,10 @@ use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 class QuestionsController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -68,7 +72,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        
+        $this->authorize("update", $question);
         return view('questions.edit', compact('question'));
     }
 
@@ -81,6 +85,7 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question);
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your Question has been updated.');
     }
@@ -92,7 +97,8 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
-    {
+    {   
+        $this->authorize("delete", $question);
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'your question has been deleted');
     }
